move code into separate module
